feature: animated text component
