+++ src/components/AnimatedText/interfaces/index.ts
@@ -1,0 +1,20 @@
+const animationStateEnum = {
+  'show-up': 'show-up',
+  'show-down': 'show-down',
+  'hide-up': 'hide-up',
+  'hide-down': 'hide-down',
+};
+
+export type AnimationState = keyof typeof animationStateEnum;
+
+export interface AnimationConfig {
+  showDelay?: string;
+  showDuration?: string;
+  hideDelay?: string;
+  hideDuration?: string;
+}
+
+export interface TextProps {
+  animationState?: AnimationState;
+  animationConfig?: AnimationConfig;
+}
