+++ src/pages/screens/Home/interfaces/index.ts
@@ -1,4 +1,4 @@
-import { AnimationState } from '@pages/HomeToPotency/HomeToPotency.styles';
+import { AnimationState } from '@components/AnimatedText/interfaces';
 
 export interface HomePageProps {
   animationState: AnimationState;
